# api-spec.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: Swapp API
  description:
    The swapp API specification. All endpoints are implicitely prefixed by __/api__ for the sake of readability.
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /items:
    get:
      description: Gets unarchived `Item` objects according to the specified filters. If no parameter are given, items that the current user could find interresting will be returned.
      parameters:
        - in: query
          name: q
          description: Query of user.
          required: false
          type: string
          
        - in: query
          name: category
          description: Category of item wanted.
          required: false
          type: string
        
        - in: query
          name: lat
          description: Latitude of current point from where to search.
          required: false
          type: number
          
        - in: query
          name: lon
          description: Longitude of current point from where to search.
          required: false
          type: number
          
        - in: query
          name: radius
          description: Radius (in kilometers) in which items must be to appear in the result.
          required: false
          type: number
          
        - in: query
          name: price_min
          description: Lower bound of an item's price range that must be respected in order for an item to appear in the result.
          required: false
          type: number
          
        - in: query
          name: price_max
          description: Upper bound of an item's price range that must be respected in order for an item to appear in the result.
          required: false
          type: number
          
        - in: query
          name: order_by
          description: Define the sorting in the result.
          required: false
          type: string
          
      responses:
        200:
          description: Successful operation.
          schema:
            type: array
            items:
              $ref: "#/definitions/DetailedItemInfo"
                  
    post:
      description: Creates a new `Item`.
      parameters:
        - in: body
          name: body
          description: The content of the item.
          required: true
          schema:
            $ref: "#/definitions/SetItemInfo"
            
      responses:
        201:
          description: Successful operation
          headers:
            Location: 
              description: Location of the created object
              type: string
  /items/{id}:
    get:
      description: Gets a detailed unarchived item.
      parameters:
        - in: path
          name: id
          description: The id of the item.
          required: true
          type: number
      responses:
        200:
          description: Successful operation
          schema:
            $ref: "#/definitions/DetailedItemInfo"
            
    put:
      description: Updates the info of the given item.
      parameters:
        - in: path
          name: id
          description: The id of the item.
          required: true
          type: number
        - in: body
          name: body
          description: The content of the item.
          required: true
          schema:
            $ref: "#/definitions/SetItemInfo"
            
      responses:
        200:
          description: Successful operation
        401:
          description: User not logged in
              
    patch:
      description: Partially updates the info of the given item.
      parameters:
        - in: path
          name: id
          description: The id of the item.
          required: true
          type: number
        - in: body
          name: body
          description: The content of the item.
          required: true
          schema:
            $ref: "#/definitions/SetItemInfo"
            
      responses:
        200:
          description: Successful operation
        401:
          description: User not logged in
...
definitions:
...
  SetItemInfo:
    type: object
    properties:
      name: 
        type: string
      description:
        type: string
      price_min:
        type: number
      price_max:
        type: number
      category:
        type: number
  DetailedItemInfo:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      description:
        type: string
      price_min:
        type: number
      price_max:
        type: number
      creation_date:
        type: string
      owner_username:
        type: string
      category:
        type: object
        properties:
          id:
            type: number
          name:
            type: string
      views:
        type: number
      image_urls:
        type: array
        items:
          type: string
      likes:
        type: number
      comments:
        type: number
      offers_received:
        type: number